backend/pyproject.toml
[project]
name = "plategallery-backend"
version = "1.0.0"
description = "PlateGallery API — FastAPI backend for vanity plate gallery"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "Pillow>=10.4.0",
    "imagehash>=4.3.0",
    "python-multipart>=0.0.12",
    "openai>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/db/migrations/**" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
